moved rolling-die component to its own folder
