+++ src/app/app.module.ts
@@ -1,17 +1,17 @@
-import { NgModule, } from '@angular/core';
-import { BrowserModule, } from '@angular/platform-browser';
+import { NgModule } from '@angular/core';
+import { BrowserModule } from '@angular/platform-browser';
 
-import { AppComponent, } from './app.component';
-import { HeaderComponent, } from './header/header.component';
-import { CharacterCreationComponent, } from './character-creation/character-creation.component';
-import { CharacterSheetComponent, } from './character-sheet/character-sheet.component';
-import { SheetComponent, } from './character-sheet/sheet/sheet.component';
-import { InfoGroupComponent, } from './character-sheet/sheet/info-group/info-group.component';
-import { DetailDisplayComponent, } from './character-sheet/sheet/detail-display/detail-display.component';
-import { DropdownDirective, } from './shared/dropdown.directive';
-import { NgbModule, } from '@ng-bootstrap/ng-bootstrap';
-import { BrowserAnimationsModule, } from '@angular/platform-browser/animations';
-import { RollingDieComponent, } from './shared/rolling-die.component';
+import { AppComponent } from './app.component';
+import { HeaderComponent } from './header/header.component';
+import { CharacterCreationComponent } from './character-creation/character-creation.component';
+import { CharacterSheetComponent } from './character-sheet/character-sheet.component';
+import { SheetComponent } from './character-sheet/sheet/sheet.component';
+import { InfoGroupComponent } from './character-sheet/sheet/info-group/info-group.component';
+import { DetailDisplayComponent } from './character-sheet/sheet/detail-display/detail-display.component';
+import { DropdownDirective } from './shared/dropdown.directive';
+import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { RollingDieComponent } from './shared/rolling-die/rolling-die.component';
 import { DieImageComponent } from './shared/die-image/die-image.component';
 import { QualityListComponent } from './shared/quality-list/quality-list.component';
 
@@ -29,8 +29,8 @@ import { QualityListComponent } from './shared/quality-list/quality-list.compone
         DieImageComponent,
         QualityListComponent,
     ],
-    imports: [BrowserModule, NgbModule, BrowserAnimationsModule,],
+    imports: [BrowserModule, NgbModule, BrowserAnimationsModule],
     providers: [],
-    bootstrap: [AppComponent,],
-},)
+    bootstrap: [AppComponent],
+})
 export class AppModule {}
