+++ src/app/shared/rolling-die/rolling-die.component.ts
@@ -1,10 +1,10 @@
-import { Component, Input, } from '@angular/core';
-import { Die, } from './die.model';
+import { Component, Input } from '@angular/core';
+import { Die } from '../die.model';
 
 @Component({
     selector: 'app-rolling-die',
     templateUrl: 'rolling-die.component.html',
-},)
+})
 export class RollingDieComponent {
     @Input() die: Die;
 }
